+++ README.md
@@ -61,6 +61,7 @@
 2. 이동 결과를 반영한 포진을 출력한다.
 3. 1~3단계를 수행하면 다음 플레이어 차례로 넘어간다.
     - 왕을 잡으면 해당 플레이어의 승리로 게임은 종료된다.
+    - 두 플레이어 모두 종료를 원할 경우, 게임은 종료된다.
 
 <br/>
