+++ src/test/java/janggi/strategy/JumpingStrategyTest.java
@@ -6,7 +6,7 @@ import janggi.direction.Direction;
 import janggi.direction.Movement;
 import janggi.direction.PieceMoveRule;
 import janggi.direction.PieceType;
-import janggi.piece.Board;
+import janggi.board.Board;
 import janggi.piece.Piece;
 import janggi.position.Position;
 import java.util.Set;
